feat: Add MetadataProcessor project and service collection extension
