+++ src/Application/Application.csproj
@@ -16,5 +16,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Engine\Engine.csproj" />
+    <ProjectReference Include="..\Features\MetadataProcessor\MetadataProcessor.csproj" />
   </ItemGroup>
 </Project>
